+++ data/duke.txt
@@ -1,3 +1,3 @@
 T/X/homework
-T/ /help
-E/X/Welcome party/25 Aug
+E/ /Welcome party/25 Aug
+D/X/iP/2021-09-05
