--- src/main/java/Command.java
@@ -1,3 +1,0 @@
-abstract public class Command {
-    public abstract void execute(Tasklist tasks, Storage storage, Ui ui) throws DukeException;
-}
