+++ src/main/java/Duke.java
@@ -1,3 +1,10 @@
+import duke.command.Command;
+import duke.error.DukeException;
+import duke.general.Parser;
+import duke.general.Storage;
+import duke.general.Tasklist;
+import duke.general.Ui;
+
 /**
  * Chatbot that helps you form a task list
  */
@@ -7,6 +14,7 @@ public class Duke {
     private Tasklist tasks;
     private Ui ui;
     private Parser parser;
+
 
     public Duke(String filePath) {
         ui = new Ui();
