--- src/main/java/ErrorCommand.java
@@ -1,6 +1,0 @@
-public class ErrorCommand extends Command {
-    @Override
-    public void execute(Tasklist tasks, Storage storage, Ui ui) {
-        System.out.println("Command error occurred, no matches.");
-    }
-}
