--- src/main/java/ListCommand.java
@@ -1,7 +1,0 @@
-public class ListCommand extends Command{
-
-    @Override
-    public void execute(Tasklist tasks, Storage storage, Ui ui) {
-        ui.showList(tasks);
-    }
-}
