+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,12 @@
+package duke.command;
+
+import duke.error.DukeException;
+import duke.general.Storage;
+import duke.general.TaskType;
+import duke.general.Tasklist;
+import duke.general.Ui;
+import duke.task.Task;
+
 public class AddCommand extends Command {
     private TaskType type;
     private String[] input;
