+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.command.Command;
+import duke.general.Storage;
+import duke.general.Tasklist;
+import duke.general.Ui;
+
 public class ByeCommand extends Command {
 
     @Override
