+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.general.Storage;
+import duke.general.Tasklist;
+import duke.general.Ui;
+import duke.task.Task;
+
 public class DeleteCommand extends Command {
     private String[] input;
