+++ src/main/java/duke/error/DukeException.java
@@ -1,9 +1,11 @@
+package duke.error;
+
 /**
  * Exceptions that can occur specifically in the Duke program.
  */
 public class DukeException extends Exception {
     /**
-     * Constructor of the DukeException
+     * Constructor of the duke.error.DukeException
      * @param message Message to be printed out to user when error has occurred.
      */
     public DukeException(String message) {
