+++ src/main/java/duke/general/Parser.java
@@ -1,3 +1,7 @@
+package duke.general;
+
+import duke.command.*;
+
 import java.util.Scanner;
 
 public class Parser {
