+++ src/main/java/duke/general/TaskType.java
@@ -1,3 +1,5 @@
+package duke.general;
+
 public enum TaskType {
     TODO, DEADLINE, EVENT
 }
